requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1
django==5.2.1
psycopg[binary,pool]==3.2.9
sqlparse==0.5.3
typing-extensions==4.12.2 ; python_version >= '3.8'
whitenoise==6.1.0
gunicorn==20.1.0
pytest==8.0.0
pytest-django==4.5.2
marshmallow>=3.0,<5.0
environs[django]==14.0,<15.0
django-cleanup==8.1.0
pillow==10.4.0
django-bootstrap5==24.3
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-imagekit==5.0.0
django-richtextfield==1.6.2
requests==2.32.3
stripe==10.4.0
django-storages==1.14.4
boto3==1.35.2
django-csp==3.8.0
django-axes==6.5.2
beautifulsoup4==4.13.4